Create and roughly style song selection component
These are placeholder textboxes for now. I will replace the top part
with a Spotify player that can search for and play songs. None of these
song selection information are being saved at all (or even part of the
entry information). The edit mode doesn't apply to them either. I'll do
that later.