Modify logic of when to continue playing/pause
The Save button and the selecting song components' logic had to be
changed so that when a song that is playing that is the same as the one
selected or saved, it continues to play. However, if the song playing
changes at some point and shouldn't continuously play, it should pause
and reset to allow whatever the user is now trying to play (some
other song) to be loaded up and played.