Apply responsive styling to calendar page
I somewhat made the calendar page responsive. I'm still not super
satisfied with it (and nothing is decorated much) because there is no
need for the calendar to scroll on large screens (like when in
fullscreen). There's also this weird jump in styling from like Laptop
to Laptop L width.